requirements.txt
Flask==2.3.2
flask-cors==4.0.0
tensorflow-cpu==2.10.0
Pillow==10.0.0
numpy==1.24.3
supabase==2.3.0
python-dotenv==1.0.0
gdown==4.7.1